Moved PassCodeManager to presentation.ui.security package
